@@ -255,7 +255,7 @@ namespace Newtonsoft.Json.Tests.Converters
             ExceptionAssert.Throws<ArgumentNullException>(
                 () => new StringEnumConverter(null),
                 @"Value cannot be null.
-Parameter name: namingStrategyType");
+Parameter name: namingStrategyType", "Value cannot be null. (Parameter 'namingStrategyType')");
         }
 
         [Test]
@@ -264,7 +264,7 @@ Parameter name: namingStrategyType");
             ExceptionAssert.Throws<ArgumentNullException>(
                 () => new StringEnumConverter(null, new object[] { true, true, true }, false),
                 @"Value cannot be null.
-Parameter name: namingStrategyType");
+Parameter name: namingStrategyType", "Value cannot be null. (Parameter 'namingStrategyType')");
         }
 
         [Test]
