@@ -84,7 +84,8 @@ namespace Newtonsoft.Json.Tests.JsonTextReaderTests
 
             await ExceptionAssert.ThrowsAsync<JsonReaderException>(
                 () => token.CreateReader().ReadAsDecimalAsync(),
-                "Could not convert to decimal: 1.79769313486232E+308. Path ''."
+                "Could not convert to decimal: 1.79769313486232E+308. Path ''.",
+                "Could not convert to decimal: 1.7976931348623157E+308. Path ''."
             );
         }
 
