@@ -1018,7 +1018,11 @@ namespace Newtonsoft.Json.Tests
                 await jsonWriter.WriteEndArrayAsync();
             }
 
+#if !NETSTANDARD2_0
             Assert.AreEqual(@"[0.0,0.0,0.1,1.0,1.000001,1E-06,4.94065645841247E-324,Infinity,-Infinity,NaN,1.7976931348623157E+308,-1.7976931348623157E+308,Infinity,-Infinity,NaN]", sb.ToString());
+#else
+            Assert.AreEqual(@"[0.0,0.0,0.1,1.0,1.000001,1E-06,5E-324,Infinity,-Infinity,NaN,1.7976931348623157E+308,-1.7976931348623157E+308,Infinity,-Infinity,NaN]", sb.ToString());
+#endif
         }
 
         [Test]
@@ -1094,7 +1098,7 @@ namespace Newtonsoft.Json.Tests
                 await ExceptionAssert.ThrowsAsync<FormatException>(async () => { await jsonWriter.WriteTokenAsync(JsonToken.Integer, "three"); }, "Input string was not in a correct format.");
 
                 await ExceptionAssert.ThrowsAsync<ArgumentNullException>(async () => { await jsonWriter.WriteTokenAsync(JsonToken.Integer); }, @"Value cannot be null.
-Parameter name: value");
+Parameter name: value", "Value cannot be null. (Parameter 'value')");
             }
         }
 
